pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whispy"
version = "0.1.0"
description = "WiFi Intelligence on ESP — CSI sensing toolkit for collection, training, visualization, and federated learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Gad Gad", email = "ggad@uwo.ca"},
]
keywords = ["wifi", "csi", "sensing", "esp32", "har", "deep-learning", "federated-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
collect = ["pyserial>=3.5"]
camera = ["opencv-python>=4.8"]
dl = ["torch>=2.0", "xgboost>=2.0"]
fl = ["flwr>=1.7"]
hf = ["huggingface_hub>=0.20"]
vis = ["matplotlib>=3.7"]
all = [
    "pyserial>=3.5",
    "opencv-python>=4.8",
    "torch>=2.0",
    "xgboost>=2.0",
    "flwr>=1.7",
    "huggingface_hub>=0.20",
    "matplotlib>=3.7",
]

[project.scripts]
whispy = "whispy.cli:main"

[project.urls]
Homepage = "https://github.com/gadgadgad/whispy"
Documentation = "https://github.com/gadgadgad/whispy#readme"

[tool.setuptools.packages.find]
include = ["whispy*"]
